only admin will delete project

+++ bms/bms_app/views.py
@@ -140,13 +140,28 @@ def project_list(request):
     return render(request, 'registration/project_list.html')
 
 def delete_project(request):
+    if request.user.is_authenticated():
+        current_user = request.user
 
     if request.method == 'POST':
         del_project = request.POST.get("del_proj_id")
+
+        check_admin = ProductUser.objects.filter(product_id=del_project,prod_user_id=current_user.id, prod_user_role_id=8)
+        print "check_admin==>",check_admin
         print "project delete id==>",del_project
-        ProductDetails.objects.filter(id=del_project).delete()
-        return HttpResponse(json.dumps({'success': True}), 
+
+        if check_admin:
+            print "ready to delete"
+            ProductDetails.objects.filter(id=del_project).delete()
+
+            return HttpResponse(json.dumps({'success': True}), 
                             content_type="application/json")
+        else:
+            print "You are not admin for this project"
+       
+            return HttpResponse(json.dumps({'success': False}), 
+                            content_type="application/json")
+
 
 @login_required(login_url='/login/')
 def create_bug(request):
